backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-pptx
chromadb
sentence-transformers
google-generativeai
python-dotenv
azure-storage-file-share
azure-storage-blob